# internal/httpapi/admin/openapi/swagger.json
{
    "components": {
        "schemas": {
...
            "admin.routeGroupInput": {
                "properties": {
                    "description": {
                        "maxLength": 500,
                        "type": "string"
                    },
                    "lcr_mode": {
                        "type": "boolean"
                    },
                    "name": {
                        "maxLength": 100,
                        "minLength": 1,
                        "type": "string"
                    }
                },
                "required": [
                    "name"
                ],
                "type": "object"
            },
...
            "model.RouteGroup": {
                "properties": {
                    "created_at": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "lcr_mode": {
                        "type": "boolean"
                    },
                    "name": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "store.RouteRow": {
                "properties": {
                    "carriers": {
                        "items": {
                            "$ref": "#/components/schemas/store.RouteCarrierRow"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "destination": {
                        "type": "string"
                    },
                    "enabled": {
                        "type": "boolean"
                    },
                    "id": {
                        "type": "string"
                    },
                    "lcr_mode": {
                        "type": "boolean"
                    },
                    "prefix": {
                        "type": "string"
                    },
                    "route_group_id": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "ApiKeyAuth": {
                "description": "\"Bearer sbc_...\" API key, or the sbc_access cookie set by POST /auth/login (with X-CSRF-Token for state changes).",
                "in": "header",
                "name": "Authorization",
                "type": "apiKey"
            }
        }
    },
    "info": {
        "description": "Session Border Controller control plane: customers, carriers, rate decks, routing, CDRs, reports.",
        "title": "SuperSBC Admin API",
        "version": "1.0"
    },
...
    "paths": {
...
        "/route-groups": {
            "get": {
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {},
                                    "type": "object"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "summary": "List route groups",
                "tags": [
                    "routing"
                ]
            },
            "post": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/admin.routeGroupInput",
                                "summary": "body",
                                "description": "route group"
                            }
                        }
                    },
                    "description": "route group",
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/model.RouteGroup"
                                }
                            }
                        },
                        "description": "Created"
                    }
                },
                "summary": "Create a route group",
                "tags": [
                    "routing"
                ]
            }
        },
        "/route-groups/{id}": {
            "delete": {
                "parameters": [
                    {
                        "description": "route group id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                },
                "summary": "Delete an unused route group",
                "tags": [
                    "routing"
                ]
            },
            "get": {
                "parameters": [
                    {
                        "description": "route group id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/model.RouteGroup"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "summary": "Get a route group",
                "tags": [
                    "routing"
                ]
            },
            "put": {
                "parameters": [
                    {
                        "description": "route group id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/admin.routeGroupInput",
                                "summary": "body",
                                "description": "route group"
                            }
                        }
                    },
                    "description": "route group",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/model.RouteGroup"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "summary": "Update a route group (name, description, LCR mode)",
                "tags": [
                    "routing"
                ]
            }
        },
        "/route-groups/{id}/routes": {
            "get": {
                "parameters": [
                    {
                        "description": "route group id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "prefix or destination",
                        "in": "query",
                        "name": "search",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/store.RouteRow"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "summary": "Routes of a group with their ordered carriers",
                "tags": [
                    "routing"
                ]
            },
            "post": {
                "parameters": [
                    {
                        "description": "route group id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/admin.routeInput",
                                "summary": "body",
                                "description": "route"
                            }
                        }
                    },
                    "description": "route",
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/store.RouteRow"
                                }
                            }
                        },
                        "description": "Created"
                    }
                },
                "summary": "Create a route (optionally with its carriers)",
                "tags": [
                    "routing"
                ]
            }
        },
        "/routes/{id}": {
            "delete": {
                "parameters": [
                    {
                        "description": "route id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                },
                "summary": "Delete a route",
                "tags": [
                    "routing"
                ]
            },
            "get": {
                "parameters": [
                    {
                        "description": "route id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/store.RouteRow"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "summary": "Get a route with its carriers",
                "tags": [
                    "routing"
                ]
            },
            "put": {
                "parameters": [
                    {
                        "description": "route id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/admin.routeInput",
                                "summary": "body",
                                "description": "route"
                            }
                        }
                    },
                    "description": "route",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/store.RouteRow"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "summary": "Update a route (and replace its carriers when given)",
                "tags": [
                    "routing"
                ]
            }
        },
        "/routes/{id}/carriers": {
            "put": {
                "parameters": [
                    {
                        "description": "route id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/admin.routeCarriersInput",
                                "summary": "body",
                                "description": "ordered carriers"
                            }
                        }
                    },
                    "description": "ordered carriers",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/store.RouteRow"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "summary": "Replace the ordered carrier list of a route (reorder endpoint)",
                "tags": [
                    "routing"
                ]
            }
        },
        "/routes/{id}/preview": {
            "get": {
                "parameters": [
                    {
                        "description": "route id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "sample number (defaults to the route prefix padded)",
                        "in": "query",
                        "name": "number",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "selling deck to compute margin against",
                        "in": "query",
                        "name": "sell_rate_group_id",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {},
                                    "type": "object"
                                }
                            }
                        },
                        "description": "OK"
                    }
                },
                "summary": "Buy rate and margin per carrier of a route for a sample number",
                "tags": [
                    "routing"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "/api/v1"
        }
    ]
}